backend/core/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.1
typing_extensions==4.13.2
whitenoise==6.9.0
